refactor: organize seed folder
