+++ packages/cli/src/lib/services/seed/index.ts
@@ -1,10 +1,3 @@
-export * from './schema';
 export * from './interfaces';
-export * from './seed-client';
-export * from './seed-loader';
-export * from './collection';
-export * from './seed-client';
-export * from './data-differ';
-export * from './data-mapper';
-export * from './id-mapper-client';
-export * from './requests';
+export * from './global';
+export * from './collections';
