+++ packages/cli/src/lib/services/seed/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './schema';
+export * from './interfaces';
